+++ lib/src/switch.dart
@@ -4,7 +4,7 @@
 import 'dart:async';
 
 import 'chain.dart';
-import 'for_map.dart';
+import 'map.dart';
 
 /// Maps events to a Stream and emits values from the most recently created
 /// Stream.
@@ -14,8 +14,8 @@ import 'for_map.dart';
 ///
 /// If the source stream is a broadcast stream, the result stream will be as
 /// well, regardless of the types of the streams produced by [map].
-StreamTransformer<S, T> switchMap<S, T>(Stream<T> map(S event)) =>
-    chain(forMap(map), switchLatest());
+StreamTransformer<S, T> switchMap<S, T>(Stream<T> convert(S event)) =>
+    chain(map(convert), switchLatest());
 
 /// Emits values from the most recently emitted Stream.
 ///