+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.0.12
 
-- Add `chain` and `forMap` for use cases where `StreamTransformer` instances are
+- Add `chain` and `map` for use cases where `StreamTransformer` instances are
   stored as variables or passed to methods other than `transform`.
 
 ## 0.0.11